[wallet/azure-hsm] fix bug when converting big number to a buffer
