+++ packages/contractkit/src/wallets/azure-hsm-wallet.ts
@@ -1,6 +1,7 @@
 import { Address, ensureLeading0x } from '@celo/utils/lib/address'
 import * as ethUtil from 'ethereumjs-util'
 import { AzureKeyVaultClient } from '../utils/azure-key-vault-client'
+import { bigNumberToBuffer } from '../utils/signing-utils'
 import { RemoteWallet } from './remote-wallet'
 import { AzureHSMSigner } from './signers/azure-hsm-signer'
 import { Signer } from './signers/signer'
@@ -51,7 +52,7 @@ export class AzureHSMWallet extends RemoteWallet implements Wallet {
       throw new Error('AzureHSMWallet needs to be initialized first')
     }
     const publicKey = await this.keyVaultClient!.getPublicKey(keyName)
-    const pkBuffer = ethUtil.toBuffer(ensureLeading0x(publicKey.toString(16)))
+    const pkBuffer = bigNumberToBuffer(publicKey, 64) // public keys are 64 bytes in size.
     if (!ethUtil.isValidPublic(pkBuffer, true)) {
       throw new Error(`Invalid secp256k1 public key for keyname ${keyName}`)
     }
